fix(notifications): raise FCM failures with the notification error codes
FCM-300/900/901/902는 정의만 되어 있고 아무도 쓰지 않았다. 정작 그 상황이 나는 자리에서는
InvalidInputException/InternalServerException을 문자열 메시지와 함께 바로 던지고 있어서,
클라이언트는 FCM 전용 코드 대신 공통 코드만 받았다. 네 코드를 실제 발생 지점에 연결한다.

- FCM_TOKEN_NOT_FOUND: 수신자 토큰 조회 실패
- FCM_INVALID_ARGUMENT / FCM_AUTH_ERROR / FCM_UNKNOWN_ERROR: FirebaseAdapter의 오류 분기
- UnregisteredTokenException을 짝이 되는 코드와 같은 파일로 옮긴다.

INVALID_ARGUMENT는 400에서 500으로 바뀐다. FCM이 이 코드를 준다는 것은 우리가 조립한 메시지가
잘못됐다는 뜻이라 클라이언트가 고칠 수 없고, 카탈로그도 이미 FCM-900을 5xx로 분류하고 있었다.
